+++ app/ready.ts
@@ -5,6 +5,9 @@ import fs from 'node:fs'
 import Discord from 'discord.js'
 import Commands from 'discord/commands'
 
+import processRoles from "./processRoles.ts"
+import processMembers from "./processMembers.ts"
+
 import Dig from './commands/stats/dig.ts'
 
 
@@ -28,15 +31,6 @@ export default async function (client: Discord.Client) {
     })
 
 
-    async function processRoles() {
-        const guild = await App.guild()
-        guild.roles.cache.forEach(r => {
-            Db.roles.updateOne({ id: r.id }, { $set: { _id: r.id, ...r.toJSON() as Role } }, { upsert: true })
-        })
-    }
-    setInterval(processRoles, 1000 * 60 * 60), processRoles()
-
-
     async function updateStatus() {
         try {
             const status = await Db.data.findOne({ _id: 'status' })
@@ -49,44 +43,9 @@ export default async function (client: Discord.Client) {
         }
         catch { }
     }
+    
+    
     setInterval(updateStatus, 1000 * 60 * 5), updateStatus()
-
-
-    async function processMembers() {
-        const guild = await App.guild()
-        await guild.members.fetch()
-        const members = guild.members.cache.filter(m => m.roles.cache.has(App.config.memberRole))
-
-        const batchSize = 10
-        const delay = ms => new Promise(res => setTimeout(res, ms))
-
-        const allMembers = Array.from(members.values())
-
-        for (let i = 0; i < allMembers.length; i += batchSize) {
-            const batch = allMembers.slice(i, i + batchSize)
-
-            // Fetch user info concurrently, but only for this batch
-            const users = await Promise.allSettled(batch.map(m => m.user.fetch()))
-
-            // Save to DB
-            for (const result of users) {
-                if (result.status === 'fulfilled') {
-                    const user = result.value
-                    let userJson = user.toJSON()
-                    userJson['guild'] = guild.members.cache.get(user.id).toJSON()
-
-                    await Db.users.updateOne(
-                        { _id: user.id },
-                        { $set: userJson },
-                        { upsert: true }
-                    )
-                }
-            }
-
-            console.log(`Processed ${i + batch.length}/${allMembers.length} members`)
-            await delay(2000)
-        }
-    }
+    setInterval(processRoles, 1000 * 60 * 60), processRoles()
     setInterval(processMembers, 1000 * 60 * 60), processMembers()
-
 }
